+++ MaterialDesignExtensions/Controls/OpenDirectoryDialog.cs
@@ -8,6 +8,7 @@ using System.Windows;
 
 using MaterialDesignThemes.Wpf;
 
+// use Pri.LongPath classes instead of System.IO for the MaterialDesignExtensions.LongPath build to support long file system paths on older Windows and .NET versions
 #if LONG_PATH
 using DirectoryInfo = Pri.LongPath.DirectoryInfo;
 #endif