+++ MaterialDesignExtensions/Converters/AsyncImageTask.cs
@@ -18,41 +18,69 @@ namespace MaterialDesignExtensions.Converters
     /// </summary>
     public class AsyncImageTask : INotifyPropertyChanged
     {
+        private readonly object m_lockObject = new object();
+
         /// <summary>
         /// The property changed event.
         /// </summary>
         public event PropertyChangedEventHandler PropertyChanged;
-        
+
+        private object m_image;
+
         /// <summary>
         /// The loaded image.
         /// </summary>
-        public object Image { get; private set; }
+        public object Image
+        {
+            get
+            {
+                lock (m_lockObject)
+                {
+                    return m_image;
+                }
+            }
+
+            private set
+            {
+                lock (m_lockObject)
+                {
+                    if (m_image != value)
+                    {
+                        m_image = value;
+
+                        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(Image)));
+                    }
+                }
+            }
+        }
 
         /// <summary>
         /// Creates a new AsyncImageTask to load the image with the specified width and height keeping its ratio.
         /// </summary>
-        /// <param name="imageFilename"></param>
-        /// <param name="targetWidth"></param>
-        /// <param name="targetHeight"></param>
+        /// <param name="imageFilename">The full filename of the image</param>
+        /// <param name="targetWidth">The target width of the image</param>
+        /// <param name="targetHeight">The target height of the image</param>
         public AsyncImageTask(string imageFilename, int targetWidth = 40, int targetHeight = 40, bool useCache = false)
         {
-            Image = PackIconKind.FileImage;
+            m_image = PackIconKind.FileImage;
 
             LoadImageAsync(imageFilename, targetWidth, targetHeight, useCache);
         }
 
         private async void LoadImageAsync(string imageFilename, int targetWidth, int targetHeight, bool useCache)
         {
-            object image = await Task.Run(() => BitmapImageHelper.LoadImage(imageFilename, targetWidth, targetHeight, useCache));
-
-            if (image == null)
+            await Task.Run(async () =>
             {
-                image = PackIconKind.Image;
-            }
+#if DEBUG
+                Console.WriteLine($"get image {imageFilename}");
+#endif
+                BitmapImage image = await BitmapImageHelper.LoadImageAsync(imageFilename, targetWidth, targetHeight, useCache).ConfigureAwait(false);
 
-            Image = image;
-
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(Image)));
+                if (image != null)
+                {
+                    Image = image;
+                }
+            }).ConfigureAwait(false);
         }
     }
 }