+++ MaterialDesignExtensions/Model/DividerNavigationItem.cs
@@ -6,8 +6,14 @@ using System.Threading.Tasks;
 
 namespace MaterialDesignExtensions.Model
 {
+    /// <summary>
+    /// A divider line in a <see cref="Controls.SideNavigation"/>.
+    /// </summary>
     public class DividerNavigationItem : NotSelectableNavigationItem
     {
+        /// <summary>
+        /// Creates a new <see cref="DividerNavigationItem" />.
+        /// </summary>
         public DividerNavigationItem() : base() { }
     }
 }