+++ MaterialDesignExtensions/Model/FileSystemEntriesGroupHeader.cs
@@ -7,6 +7,9 @@ using System.Threading.Tasks;
 
 namespace MaterialDesignExtensions.Model
 {
+    /// <summary>
+    /// A group header in the items lists of file system controls to group directories and files.
+    /// </summary>
     public class FileSystemEntriesGroupHeader : INotifyPropertyChanged
     {
         public event PropertyChangedEventHandler PropertyChanged;
@@ -14,6 +17,9 @@ namespace MaterialDesignExtensions.Model
         private string m_header;
         private bool m_showSeparator;
 
+        /// <summary>
+        /// The header label.
+        /// </summary>
         public string Header
         {
             get
@@ -32,6 +38,9 @@ namespace MaterialDesignExtensions.Model
             }
         }
 
+        /// <summary>
+        /// True, to display a separator above the label.
+        /// </summary>
         public bool ShowSeparator
         {
             get
@@ -50,6 +59,9 @@ namespace MaterialDesignExtensions.Model
             }
         }
 
+        /// <summary>
+        /// Creates a new <see cref="FileSystemEntriesGroupHeader" />.
+        /// </summary>
         public FileSystemEntriesGroupHeader()
         {
             m_header = null;