+++ MaterialDesignExtensions/Model/NavigationItem.cs
@@ -6,11 +6,17 @@ using System.Threading.Tasks;
 
 namespace MaterialDesignExtensions.Model
 {
+    /// <summary>
+    /// An navigation item of a <see cref="Controls.SideNavigation" />.
+    /// </summary>
     public class NavigationItem : BaseNavigationItem
     {
         private object m_icon;
         private string m_label;
 
+        /// <summary>
+        /// The label of this navigation item.
+        /// </summary>
         public object Icon
         {
             get
@@ -26,6 +32,9 @@ namespace MaterialDesignExtensions.Model
             }
         }
 
+        /// <summary>
+        /// The label of this navigation item.
+        /// </summary>
         public string Label
         {
             get
@@ -41,6 +50,9 @@ namespace MaterialDesignExtensions.Model
             }
         }
 
+        /// <summary>
+        /// Creates a new <see cref="NavigationItem" />.
+        /// </summary>
         public NavigationItem()
         {
             m_icon = null;